Refactor sequential and parallel implementation into two files
